Restructure ACE moving the by_point in utils and renameing ace to ace_by_track
